chibiOS modifications to enable fast ADC sampling
The ADC abstraction layer was gutted to reduce a bunch of
the redundant checks for this use case.

now, the system can only handle one channel, and assumes cyclic
conversion. The overhead of doing all the checks for everything else
every sample was too high. The code is a lot less safe now, as
in you can bork it by nuking the ADC group config records, but
it runs faster. Tread with caution.