toying with an optimization on systick to avoid missing samples
The code in this commit was eventually #if'd out, but wanted to
remind myself that there could be a way to catch missed samples
by instrumenting the systick handler to check and see if an
interrupt fired while we were away, and taking care of it.

I /believe/ the stack masks all other interrupts while in a context,
so if the ADC triggers while systick goes off, we miss a sample.