add updater protocol
This is a protocol that runs on top of the MAC layer which
processes multiple packets out of the MAC and updates the
application code space accordingly.

Currently the protocol doc resides in updater.c. Will need
to add some .md file describing it in case people want to
develop their own interoporable programming front ends.