app: add app support
Apps are chunks of data with predefined headers.  Add app support, along with
code to launch the apps.  The program's data section is copied, and its BSS
is cleared.  Any constructors are called.

Note that the laucher does not disable interrupts or do anything like that,
so you will need to do that prior to calling Run_App().

Signed-off-by: Sean Cross <xobs@kosagi.com>