KINETIS: adc: add hooks for ISR, fast ISR
We want to override the ISR sometimes.  Add hooks to allow us to do this.

There are two hooks: The first one is a "fast" ISR, which will override
the ISR before it gets to ChibiOS.

The "normal" ISR will allow you to still call ChibiOS functions from
within the handler.

Signed-off-by: Sean Cross <xobs@kosagi.com>