svc_handler: load halfword elements from svc_handlers
Since we've moved to 16-bit SVC handler addresses, we've needed to
modify the SVC handler to load 16-bit values.  All of this was done,
except we were still using "ldr" instead of "ldrh", which segfaults on
non-32-bit offsets.

Change the "ldr" to "ldrsh" to load the 16-bit value into the 32-bit
register, and clear the upper bits.

Signed-off-by: Sean Cross <xobs@kosagi.com>