fix time offset bug
If users are developing code for a while and the sticker is
waiting for the code to be uploaded, then application time
starts with an offset.

Arduino assumes time starts at 0 so you can just read
millis() and get the time offset since application launch.

Add code which handles this offset so that applications launch
with the correct notion of time since launch even if the
sticker is idle for a while before receiving the app.