orchard: add option to disable ISR lock
Normally, you want to disable IRQs when entering a critical section in
an IRQ handler.  Therefore, various functions call chSysLockFromISR() and
chSysUnlockFromISR().

However, we have set up priorities so that the vast majority of our interrupts
are on the same priority level, and those that aren't are re-entrant.

Furthermore, we have timing-critical code that needs to be able to be
interrupted at any time.

Ordinarily we'd use an NMI, but the bootstrapping problem is not a thing
we'd like to tackle right now.

So as a result, don't disable interrupts during ISRs.  This produces good
results, and has not led to instability.

Signed-off-by: Sean Cross <sean@xobs.io>