orchard-vectors: jump to errorCondition rather than bkpt
Now that we're getting ready for a release, have unhandled faults call
errorCondition(), which blinks a light.

Note that on the M0, there is no way to determine if a debugger is
attached.  Therefore, build with -DNDEBUG, which will enable the call to
bkpt in order to break into the debugger.

Signed-off-by: Sean Cross <sean@xobs.io>