add a debug snippet for checking signal quality of analog demod
When this code snippet is set to be included, it causes PTA12 (LED 2
on PVT, LED A1 on DVT) to flash "high" when a high tone is detected,
and "low" when a low tone is detected. This only happens during
upload mode.

Very useful for debugging demodulator/audio quality problems.