Makefile: temporarily lower exceptions stacksize by 8 bytes
The new GCC is fragmenting memory.  It ought to pack u8 values together, but
instead it's grouping accesses by file.

For example, sudo_mode is a u8, and pwm_status is a u8.  However, because
each is in its own file, they are padded with three bytes after each,
wasting six bytes in total.

It is unclear how to turn this off.  May need to move to using a constant
pool instead.

Signed-off-by: Sean Cross <sean@xobs.io>