boards: ltc: reset usb lines during __early_init()
Currently, we set the USB lines during board_init().  This is early
on during boot, but not early enough.

It could be that crystal calibration takes too long, or some other
process causes the clock initialization to be delayed to the point
where the host system attempts to enumerate a USB device.

Avoid all of this by driving the USB lines low before the clock is
calibrated.

Signed-off-by: Sean Cross <sean@xobs.io>